// src/app/pages/system-setup/company-setup/models/company-setup.model.ts
export class CompanySetup {
    address: string;
    website: string;
    email: string;
    phoneNumber: String;

    avayaCentral: boolean;
    avayaAPIKeyInformation: string;
    avayaAPISecretKey: string;

    twilioSendGrid: boolean;
    twilioAPIKeyInformation: string;
    twilioAPISecretKey: string;

    twilioWhatsApp: boolean;
    twilioWhatsAppAPIKeyInformation: string;
    twilioWhatsAppAPISecretKey: string;

    SMSProxireach: boolean;
    SMSProxireachAPIKeyInformation: string;
    SMSProxireachAPISecretKey: string;
}